% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powergene.R
\docType{package}
\name{powergene-package}
\alias{powergene}
\alias{powergene-package}
\title{powergene: A Simulation-Based Power Analysis for Semi-Field Vector Control Trials}
\description{
The methods for a simulation-based power analysis tailored to short- or long-term semi-field vector control trials testing single or combined interventions.
The \code{powergene} package supports random effects modeling, empirical p-value extraction, and
parallelized workflows. It is designed to guide experimental design and improve statistical rigor
in entomological research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/akipingu/powergene}
  \item Report bugs at \url{https://github.com/akipingu/powergene/issues}
}

}
\author{
\strong{Maintainer}: Andrea M. Kipingu \email{akipingu@ihi.or.tz} (\href{https://orcid.org/0000-0002-9063-8651}{ORCID}) (Department of Environmental Health & Ecological Sciences, Ifakara Health Institute, P.O. Box 78 373, Dar es Salaam, Tanzania)

Authors:
\itemize{
  \item Samson S. Kiware \email{skiware@ihi.or.tz} (\href{https://orcid.org/0000-0001-7252-520X}{ORCID}) (Department of Environmental Health & Ecological Sciences, Ifakara Health Institute, P.O. Box 78 373, Dar es Salaam, Tanzania)
}

}
