% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Power_Analysis_ShortTermSFE_CombinedIntervention.R
\name{sim.plot.shortsfe.comint}
\alias{sim.plot.shortsfe.comint}
\title{Plot Mosquito Counts for Short-Term Semi-Field Experiment Testing Combined Interventions}
\usage{
sim.plot.shortsfe.comint(
  n.ch.per.trt,
  lambda,
  intvn1.effect,
  intvn2.effect,
  ixn.effect,
  chamber.var,
  use.random = TRUE,
  jitter = TRUE
)
}
\arguments{
\item{n.ch.per.trt}{Integer. Number of chambers per Treatments combination.}

\item{lambda}{Numeric. Mean mosquito count in control chambers.}

\item{intvn1.effect}{Numeric. Proportion reduction due to intervention 1.}

\item{intvn2.effect}{Numeric. Proportion reduction due to intervention 2.}

\item{ixn.effect}{Numeric. Proportion reduction due to interaction between interventions.}

\item{chamber.var}{Numeric. Variance of random chamber effects.}

\item{use.random}{Logical or NULL.
If \code{TRUE}, plots mosquito counts simulated with random effects;
If \code{FALSE}, plots counts based on fixed effects only;
If \code{NULL}, plots expected counts from fixed effects (no sampling).}

\item{jitter}{Logical. If \code{TRUE}, overlays individual chamber-level points.}
}
\value{
A `ggplot` object showing mosquito counts by Treatments combination.
}
\description{
Generates a boxplot of simulated mosquito counts grouped by intervention combinations.
Uses output from `sim.mosq.shortsfe.comint()` and overlays jittered points to show chamber-level variation.
}
\examples{
sim.plot.shortsfe.comint(
  n.ch.per.trt = 4,
  lambda = 50,
  intvn1.effect = 0.7,
  intvn2.effect = 0.8,
  ixn.effect = 0.5,
  chamber.var = 0.1807,
  use.random = TRUE,
  jitter = TRUE
)

}
