% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Power_Analysis_ShortTermSFE_CombinedIntervention.R
\name{sim.pval.shortsfe.comint}
\alias{sim.pval.shortsfe.comint}
\title{Extract p-value from Simulated GLMM for Short-Term Semi-Field Experiment Testing Combined Interventions}
\usage{
sim.pval.shortsfe.comint(
  n.ch.per.trt,
  lambda,
  intvn1.effect,
  intvn2.effect,
  ixn.effect,
  chamber.var,
  use.random = TRUE
)
}
\arguments{
\item{n.ch.per.trt}{Integer. Number of chambers per Treatments combination.}

\item{lambda}{Numeric. Mean mosquito count in control chambers.}

\item{intvn1.effect}{Numeric. Proportion reduction due to intervention 1.}

\item{intvn2.effect}{Numeric. Proportion reduction due to intervention 2.}

\item{ixn.effect}{Numeric. Proportion reduction due to interaction between interventions.}

\item{chamber.var}{Numeric. Variance of random chamber effects.}

\item{use.random}{Logical.
If \code{TRUE}, uses mosquito counts simulated with random effects;
If \code{FALSE}, uses counts based on fixed effects only.}
}
\value{
A named numeric vector:
\describe{
  \item{pvalue}{p-value from likelihood ratio test comparing models with and without the interaction term}
}
}
\description{
Simulates mosquito count data under a factorial intervention design, fits two generalized linear mixed models (GLMMs),
and extracts the p-value for the interaction term via likelihood ratio test. This function supports empirical power analysis
by quantifying the statistical significance of combined intervention effects.
}
\examples{
sim.pval.shortsfe.comint(
  n.ch.per.trt = 4,
  lambda = 50,
  intvn1.effect = 0.7,
  intvn2.effect = 0.8,
  ixn.effect = 0.5,
  chamber.var = 0.1807,
  use.random = TRUE
)

}
