% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Power_Analysis_ShortTermSFE_SingleIntervention.R
\name{sim.power.shortsfe.sinint}
\alias{sim.power.shortsfe.sinint}
\title{Estimate Empirical Power for Short-Term Semi-Field Intervention}
\usage{
sim.power.shortsfe.sinint(
  nsim,
  n.ch.per.trt,
  lambda,
  interv.effect,
  chamber.var,
  n.cores = parallel::detectCores() - 1
)
}
\arguments{
\item{nsim}{Integer. Number of simulation replicates.}

\item{n.ch.per.trt}{Integer. Number of chambers per treatment level.}

\item{lambda}{Numeric. Mean mosquito count in control chambers.}

\item{interv.effect}{Numeric. Proportion reduction due to ITN.}

\item{chamber.var}{Numeric. Variance of random chamber effects.}

\item{n.cores}{Integer. Number of cores to use for parallel processing.}
}
\value{
A named numeric vector:
\describe{
  \item{power}{Estimated empirical power (rounded to two decimal places)}
}
}
\description{
Runs repeated simulations and GLMM fits to estimate empirical power
as the proportion of simulations with p-values below 0.05.
}
\examples{
sim.power.shortsfe.sinint(100, 4, 50, 0.8, 0.1807)

}
